Task1/requirements.txt
requests
beautifulsoup4
newspaper3k
python-dotenv
streamlit
lxml_html_clean
